arm: rockchip: Drop GRF_GPIOXX_IOMUX enums from RK3128's GRF header
Those enums are not used anywhere for RK3128 and they will cause problems
if both RK3128's and RK3288's grf header is included in a driver, because
RK3288's grf header defines enums with members which have exactly the same
name as these.
If they are required, they can be defined inline as it is done for most
other Rockchip SoCs.

Signed-off-by: Alex Bee <knaerzche@gmail.com>